Make sure collections exist before deleting them in migrations

+++ server/migrations/20220622085143-drop-useless-collection.js
@@ -1,8 +1,14 @@
 module.exports = {
   async up(db) {
-    db.collection("cfasAnnuaire").drop();
-    db.collection("croisementCfaAnnuaireassociés").drop();
-    db.collection("croisementVoeuxAffelnet").drop();
+    if (db.collections.cfasAnnuaire) {
+      db.collection("cfasAnnuaire").drop();
+    }
+    if (db.collections.croisementCfaAnnuaireassociés) {
+      db.collection("croisementCfaAnnuaireassociés").drop();
+    }
+    if (db.collections.croisementVoeuxAffelnet) {
+      db.collection("croisementVoeuxAffelnet").drop();
+    }
   },
   async down() {},
 };
